Refactored clientComponents into separate component files
